refactor(alerts): share missing alert modal
Generated-By: PostHog Desktop
Task-Id: e69d84a3-1e55-483f-b640-2d6b930c5ce4
