+++ products/alerts/frontend/views/EditAlertModal.tsx
@@ -2,13 +2,9 @@ import { useActions, useValues } from 'kea'
 import { Form } from 'kea-forms'
 import { useCallback, useMemo } from 'react'
 
-import * as magnifyingGlassPng from '@posthog/brand/hoggies/png/magnifying-glass-1'
-
-import { pngHoggie } from 'lib/brand/hoggies'
 import { UserActivityIndicator } from 'lib/components/UserActivityIndicator/UserActivityIndicator'
 import { dayjs } from 'lib/dayjs'
 import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
-import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { LemonModal } from 'lib/lemon-ui/LemonModal'
 import { formatDate } from 'lib/utils/datetime'
 import { insightLogic } from 'scenes/insights/insightLogic'
@@ -49,10 +45,9 @@ import { supportsAnomalyDetection, supportsOngoingInterval } from '../types'
 import type { AlertType } from '../types'
 import { AlertHistorySection } from './AlertHistorySection'
 import { AlertEnabledAction, AlertLeadingActions } from './EditAlertModal/AlertLeadingActions'
+import { AlertNotFoundModal } from './EditAlertModal/AlertNotFoundModal'
 import { buildWizardSteps } from './EditAlertModal/buildWizardSteps'
 import { defaultAlertTabs, EditAlertTabs } from './EditAlertModal/EditAlertTabs'
-
-const HedgehogMagnifyingGlass = pngHoggie(magnifyingGlassPng)
 
 interface AlertModalCommonProps {
     isOpen: boolean | undefined
@@ -415,19 +410,14 @@ export function EditAlertModal(props: AlertModalProps): JSX.Element {
     const nameError = alertFormSubmitAttempted && !alertForm.name ? 'Enter an alert name.' : undefined
     const scheduleRestrictionFormError = quietHoursFormError(alertForm.schedule_restriction)
 
+    if (alertId && !alertLoading && !alert) {
+        return <AlertNotFoundModal isOpen={Boolean(isOpen)} onClose={handleClose} />
+    }
+
     return (
         <LemonModal onClose={handleClose} isOpen={isOpen} width={900} simple title="">
             {alertLoading && !alert ? (
                 <AlertEditorLoading title="Edit alert" onBack={handleClose} />
-            ) : alertId && !alert ? (
-                <div className="flex min-h-[600px] flex-col items-center justify-center gap-3 p-6 text-center">
-                    <HedgehogMagnifyingGlass className="h-28 w-28" />
-                    <h2 className="m-0 text-lg font-semibold">Alert not found</h2>
-                    <p className="m-0 text-secondary">This alert may have been deleted.</p>
-                    <LemonButton type="secondary" onClick={handleClose}>
-                        Back to alerts
-                    </LemonButton>
-                </div>
             ) : (
                 <Form
                     logic={alertFormLogic}
